Responses marks are calculated on the server
